requirements.txt
deepface
ultralytics
fiftyone
pymongo
transformers[torch]
timm
torchvision
Pillow
numpy
datasets
accelerate
albumentations
torchmetrics
rfdetr